+++ src/widgets/hero/model/heroConfig.ts
@@ -1,0 +1,3 @@
+export const showScrollArrow =
+  import.meta.env.MODE === 'production' ||
+  import.meta.env.VITE_SHOW_DOWN_ICON === 'true';
