+++ src/widgets/hero/styles/Hero.css
@@ -130,7 +130,6 @@
     height: 37.5vmin;
     width: auto;
     filter: none;
-    margin: 0;
   }
 
   .hero__company-text {
